фиксы тестов WIP; удалена костыльная обработка типов колонок; автолоукейс простых идентификаторов

+++ apgdiff.tests/src/main/resources/cz/startnet/utils/pgdiff/add_column_null_diff.sql
@@ -11,4 +11,4 @@ ALTER TABLE testtable
 	ALTER COLUMN field6 SET DEFAULT 0;
 
 ALTER TABLE testtable
-	ADD COLUMN field7 double DEFAULT 1 NOT NULL;
+	ADD COLUMN field7 double precision DEFAULT 1 NOT NULL;
